update(aws): Disable aws auth integration by default
AWS integration requires an account to be setup, and bloats the build
from 243 to 1200+ files. An impedement to quick and dirty evaluation of
the stack. A couple of lines commented out in core.cljs which can be
reinstated to use aws auth if it's relevant.

Dependencies still listed in package.json for now.